Move login out of classes
