+++ core/discotron.js
@@ -5,7 +5,7 @@ const Plugin = require("./classes/plugin.js");
 const Owner = require("./classes/owner.js");
 const SpamUser = require("./classes/spam-user.js");
 const Logger = require("./utils/logger.js");
-const Login = require("./classes/login.js");
+const Login = require("./login.js");
 const db = require("./database/crud.js");
 const discordClientProvider = require("./apis/discord-client-provider.js");
