+++ core/login.js
@@ -1,12 +1,13 @@
 /**
  * Handle user login and first login
  */
-const webAPI = require("../../dashboard/backend/api.js").getWebAPI("discotron-dashboard");
-const Owner = require("./owner.js");
 const uuidv1 = require("uuid/v1");
 const request = require("request");
-const Logger = require("../utils/logger.js");
-const db = require("../database/crud.js");
+
+const webAPI = require("../dashboard/backend/api.js").getWebAPI("discotron-dashboard");
+const Owner = require("./classes/owner.js");
+const Logger = require("./utils/logger.js");
+const db = require("./database/crud.js");
 
 const appConfig = require(global.discotronConfigPath + "/bot.json");
